+++ src/riscv.rs
@@ -4,7 +4,7 @@
 macro_rules! reg {
     ( $r:ident ) => {
         {
-            let value: usize;
+            let value: u64;
             #[allow(unused_unsafe)]
             unsafe { asm!(concat!("mv $0, ", stringify!($r)) : "=r"(value)) };
             value
@@ -16,7 +16,7 @@ macro_rules! reg {
 macro_rules! csrr {
     ( $r:ident ) => {
         {
-            let value: usize;
+            let value: u64;
             #[allow(unused_unsafe)]
             unsafe { asm!("csrr $0, $1" : "=r"(value) : "i"(crate::csr::$r)) };
             value
@@ -28,7 +28,7 @@ macro_rules! csrr {
 macro_rules! csrw {
     ( $r:ident, $x:expr ) => {
         {
-            let x: usize = $x;
+            let x: u64 = $x;
             #[allow(unused_unsafe)]
             unsafe { asm!("csrw $0, $1" :: "i"(crate::csr::$r), "r"(x)) };
         }
@@ -39,7 +39,7 @@ macro_rules! csrw {
 macro_rules! csrwi {
     ( $r:ident, $x:expr ) => {
         {
-            const X: usize = $x;
+            const X: u64 = $x;
             #[allow(unused_unsafe)]
             unsafe { asm!("li t0, $1
                             csrw $0, t0"
@@ -55,7 +55,7 @@ macro_rules! csrwi {
 macro_rules! csrs {
     ( $r:ident, $x:expr ) => {
         {
-            let x: usize = $x;
+            let x: u64 = $x;
             #[allow(unused_unsafe)]
             unsafe { asm!("csrs $0, $1" :: "i"(crate::csr::$r), "r"(x)) };
         }
@@ -66,7 +66,7 @@ macro_rules! csrs {
 macro_rules! csrsi {
     ( $r:ident, $x:expr ) => {
         {
-            const X: usize = $x;
+            const X: u64 = $x;
             #[allow(unused_unsafe)]
             unsafe { asm!("li t0, $1
                            csrs $0, t0"
@@ -82,7 +82,7 @@ macro_rules! csrsi {
 macro_rules! csrc {
     ( $r:ident, $x:expr ) => {
         {
-            let x: usize = $x;
+            let x: u64 = $x;
             #[allow(unused_unsafe)]
             unsafe { asm!("csrc $0, $1" :: "i"(crate::csr::$r), "r"(x)) };
         }
@@ -93,7 +93,7 @@ macro_rules! csrc {
 // macro_rules! csrci {
 //     ( $r:ident, $x:expr ) => {
 //         {
-//             const X: usize = $x;
+//             const X: u64 = $x;
 //             #[allow(unused_unsafe)]
 //             unsafe { asm!("csrci $0, $1" :: "i"(crate::csr::$r), "i"(X)) };
 //         }
